Filter speech/music items parsed
The basic usage is to strip down resources that aren't required by the demo version. There is also an isolated instance where the source 'music.clu' had a different version of a music file using the same ID and I had to overwrite it with an external WAV from the demo instead.